refactor: stub for useAnimatedKeyboardHandler on web
