+++ src/reanimated.web.ts
@@ -1,0 +1,1 @@
+export const useAnimatedKeyboardHandler = () => null;
